# SslSocket::shutdown logs
# VMS-5018 add multicast warning
#CLOUD-793 Added nx::utils::move_if: STL-like routine which moves values satisfying criteria from one range to another.
#CLOUD-793 Added test which reproduces the problem.
#CLOUD-793 Coding style fixes
#CLOUD-793 Fixed: aio removes posted functions with no mutex locked.
#CLOUD-793 Rename: AioTaskQueue::addSocketTask -> AioTaskQueue::addTask
#CLOUD-931 Fixed typo
#CLOUD-931 RandomOnlineEndpointSelector test made independent of Internet connection
#CLOUD-941 Added generation of cloudModules.xml based on template
#VMS-5497 Fixed race condition in test
#VMS-5565 Added more logs
Add language selector to lozalised files
Added -Werror switch to some projects
Added TODO
Bootstrapper - detect language
Coding style fixes
Coding style fixes. Some AIOThreadImpl members made private (where possible without refactoring)
Enabled -Werror on cloud_connectivity_ut
Enabled cloud_connect_test_util build on nx1
First two languages
Fix header styles
Fix language selector
Fix language.json
Fixed client crash when hovering over bookmarks
Fixed compile error
Fixed compile warning
Fixed typo
Fixed warning
Fixed: wrong result code in mediator logs
Fun tests old: disable obsolete tests
Language selector
Minor coding style fix.
Project list file for 2.6
Renaming: AioThreadImpl->AioTaskQueue
Senturian iphone build
Socket implementation code made a bit more clear.
Try to fix 'Bad substitution'
UT-42
Use language list from config.js
VMS-1861 Send statistic after updating system
VMS-5245: Small fix for Cellinx ptz mapping
VMS-5337: Add Rest API breaking changes info to web-admin
VMS-5493: Setup Wizard must be translatable
VMS-5512: Screenshot has to be in hi-res
VMS-5530: Export Audio in nov doesn't work
VMS-5541 - Crash detected in client: f67501de98221a023868bf0e9217e09b
VMS-5542 Axis speakers don't work
VMS-5556 Rtsp stream works in vlc not in Nx
VMS-5557 NX1 crash at stop (build 3.0.0.14114) - libc_do_syscall
VMS-5557 fix
VMS-5559 set ulimit file-max doesn't work
VMS-5567: Vista camera tilt control fix
VMS-5577 NIC can be disabled during hardware id calculation
VMS-5578: TestCamera - support separate camera-files
VMS-5587: Couldn't bound system to Cloud because /api/cloud_modules.xml is blocked by some other software
VMS-5591:Video layout fix for AV20365
VMS-5597 Hikvision DS-2CD6362F-I stream profiles are attached to the channels incorrectly
VMS-5600: License failover period should be 30 days
VMS-5601: NX1 mediaserver crash with signal 11, Segmentation fault ( FrameTypeExtractor::getFrameType)
WEB-475: 'Merge systems' header is visible for user who can't merge
WEB-477: Set port value to '1' -> the site can't be reached
WEB-479: Web admin language selector fixes
generate language.json
hotfix - misspelling
ios signing for senturian
turkish translations-take 2
vms-4405. Backout.
vms-4405. cdb result code UT refactored.
vms-5571. Measure archive cleanup speed.
vms-5589. Don't select storages with free space below 150 mb for recording.
vms-5609. Removed redundant warnings.

<==== Mercurial data ====>
Branch: vms_3.1_dev

+++ unit_tests/cloud_connectivity_ut/src/tunnel/incoming_tunnel_udt_connection_ut.cpp
@@ -193,7 +193,7 @@ TEST_F(IncomingTunnelConnectionTest, SynAck)
             [this, &buffer, &promise](SystemError::ErrorCode code, size_t size)
             {
                 ASSERT_EQ(code, SystemError::noError);
-                ASSERT_EQ(size, buffer.size());
+                ASSERT_EQ(size, (size_t)buffer.size());
 
                 buffer.resize(0);
                 m_freeSocket->readSomeAsync(
@@ -238,7 +238,7 @@ TEST_F(IncomingTunnelConnectionTest, SynAck)
             [this, &buffer, &promise](SystemError::ErrorCode code, size_t size)
             {
                 ASSERT_EQ(code, SystemError::noError);
-                ASSERT_EQ(size, buffer.size());
+                ASSERT_EQ(size, (size_t)buffer.size());
 
                 buffer.resize(0);
                 buffer.reserve(1);
